requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# HTTP Client
requests==2.31.0

# Environment and Config
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0

# Logging
python-json-logger==2.0.7

# Google Cloud (opcional, se usar recursos GCP)
google-cloud-logging==3.8.0

# Testing (desenvolvimento)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # Para testes de FastAPI

# Monitoring (opcional)
prometheus-client==0.19.0